pyproject.toml
[tool.poetry]
name = "emt"
version = "0.1.0"
description = ""
authors = ["Anatoly Gusev <a.gusev@sparklingtide.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
django-cors-headers = "^3.11.0"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
djangorestframework-jwt = "^1.11.0"
factory-boy = "^3.2.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
sentry-sdk = "^1.5.7"
whitenoise = "^6.0.0"
django-s3-storage = "^0.13.6"
drf-yasg = {git = "https://github.com/PetrDlouhy/drf-yasg.git", rev = "0c14a4fa905cec9875d64f36fd3bfe4032528b08"}
django-anymail = {extras = ["mailgun"], version = "^8.5"}
django-extensions = "^3.1.5"
pytest = "^7.1.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.7.0"
pytest-factoryboy = "^2.1.0"
flake8 = "^4.0.1"
hvac = "^0.11.2"
transliterate = "^1.10.2"
django-polymorphic = "^3.1.0"
djangorestframework-simplejwt = "^5.1.0"
django-mptt = "^0.13.4"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pre-commit = "^2.17.0"
isort = "^5.10.1"
rope = "^0.23.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
